classes: skarnet: fix sysdepdir paths
Defining paths using STAGING_DIR variables did not work in all cases,
because in the native case, '${libdir}' is already containing
STAGING_DIR_NATIVE in its path and, therefore, making the path wrong
when trying to import the sysdeps folder.
To make it work in all cases, set sysdepdir in the class and do not make
it user-configurable to avoid errors.

Signed-off-by: Mathieu Anquetin <mathieu@anquetin.eu>