s6: add recipe
s6 is skarnet.org's small and secure supervision suite, a suite of
programs for UNIX designed to allow process supervision in the line of
daemontools and runit. It is meant to be a toolbox for low-level process
and service administration.

For example, s6 can provide the following services by assembling several
programs together:
  - syslogd functionality
  - service readiness notification
  - controlled privileged gain without suid programs
  - socket activation

Signed-off-by: Mathieu Anquetin <mathieu@anquetin.eu>