execline: add nsss to PACKAGECONFIG
Some execline utilities (like homeof) require access to the
user/group/shadow database. This access can be provided by the system
libc or by the nsss utility provided by this layer. Therefore, this
decision is left to the distribution maintainers using a PACKAGECONFIG
option to select it at build time.

Signed-off-by: Mathieu Anquetin <mathieu@anquetin.eu>