utmps: add recipe
utmps is an implementation of the utmpx.h family of functions performing
user accounting on UNIX systems. Traditionally, this functionality is
provided by the system's libc. However, not all libcs implement utmp
functionality because it is difficult to implement it in a secure way.

utmps is a secure implementation of user accounting using a daemon as
the only authority to manage the utmp/wtmp data. Therefore, programs
running utmp functions are just clients to this daemon.

Signed-off-by: Mathieu Anquetin <mathieu@anquetin.eu>