mdevd: add recipe
mdevd is a small daemon managing kernel hotplug events, similarly to
udevd. It uses the same configuration file as mdev, the hotplug manager
integrated in busybox. However, unlike mdev, it does not fork. It can
also be used as a lightweight replacement for busybox mdev because it
does not pull any unnecessary code when busybox functinalities are
provided by alternatives.

Signed-off-by: Mathieu Anquetin <mathieu@anquetin.eu>